+++ app/services/book.service.js
@@ -5,61 +5,37 @@ import PublisherRepository from "../repositories/publisher.repository.js";
 import MongoDB from "../utils/mongodb.util.js";
 import BookCopyService from "./bookCopy.service.js";
 import { deleteFromCloudinary, uploadFromBuffer } from "./cloudinary.service.js";
-import { Status } from "./../enums/status.enum.js";
 import { BookCopyStatus } from "../enums/bookCopyStatus.enum.js";
+import LoanDetailRepository from "../repositories/loanDetail.repository.js";
 class BookService {
     constructor() {
         this.bookRepository = new BookRepository(MongoDB.client);
         this.categoryRepository = new CategoryRepository(MongoDB.client);
         this.authorRepository = new AuthorRepository(MongoDB.client);
         this.publisherRepository = new PublisherRepository(MongoDB.client);
-        this.bookCopyService = new BookCopyService();
+        this.bookCopyService = new BookCopyService(MongoDB.client);
+        this.loanDetailRepository = new LoanDetailRepository(MongoDB.client);
     }
 
     async create(bookData, files) {
-
-        if (bookData.id == "" && (bookData.name == "" || bookData.price == null || bookData.publisherId == null
-            || bookData.categoryIds.length == 0 || bookData.authorIds.length == 0 || bookData.bookCopyQuantity == "")) {
-            throw new ApiError(400, "Params not valid")
-        }
         var imageUrls = [];
-        if (bookData.images && bookData.images.length > 0) {
-            if (Array.isArray(bookData.images)) {
-                imageUrls.push(...bookData.images)
-            }
-            else {
-                imageUrls.push(bookData.images);
-            }
-        }
-        if (files || files.length != 0) {
+        if (files && files.length != 0) {
             const uploadPromises = files.map(file => uploadFromBuffer(file));
             imageUrls.push(...(await Promise.all(uploadPromises)));
         }
-
         var dataToSave = {
             ...bookData,
             images: imageUrls
         }
-        var categories = []
-        if (Array.isArray(bookData.categoryIds)) {
-            categories = await Promise.all(
-                bookData.categoryIds.map(id => this.categoryRepository.findById(id))
-            );
-        }
-        else {
-            categories.push(await this.categoryRepository.findById(bookData.categoryIds));
-        }
+        const categories = await Promise.all(
+            bookData.categoryIds.map(id => this.categoryRepository.findById(id))
+        );
 
-        var authors = [];
-        if (Array.isArray(bookData.authorIds)) {
-            authors = await Promise.all(
-                bookData.authorIds.map(id => this.authorRepository.findById(id))
-            );
-        }
-        else {
-            authors.push(await this.authorRepository.findById(bookData.authorIds));
-        }
-        var publisher = await this.publisherRepository.findById(bookData.publisherId);
+        const authors = await Promise.all(
+            bookData.authorIds.map(id => this.authorRepository.findById(id))
+        );
+
+        const publisher = await this.publisherRepository.findById(bookData.publisherId);
 
         dataToSave.categories = categories;
         dataToSave.authors = authors;
@@ -73,13 +49,84 @@ class BookService {
         return newBook;
     }
 
+    async update(bookData, files) {
+        var imageUrls = [];
+        const oldBook = await this.bookRepository.findById(bookData.id);
+        if (bookData.images && (oldBook.images.length != bookData.images.length)) {
+            const imageOldDeletes = oldBook.images.filter((item) =>
+                !bookData.images.includes(item)
+            )
+            const deletePromises = imageOldDeletes.map(element => {
+                var start = element.indexOf("book-lending-project");
+                const public_id = element.substring(start, element.lastIndexOf('.'));
+                console.log(public_id);
 
-    async findAll({ page = 1, limit = 10 }) {
-        const books = await this.bookRepository.findAll(page, limit);
+                return deleteFromCloudinary(public_id);
+            });
+            await Promise.all(deletePromises);
+        }
+        else if (!bookData.images) {
+            const deletePromises = oldBook.images.map(element => {
+                var start = element.indexOf("book-lending-project");
+                const public_id = element.substring(start, element.lastIndexOf('.'));
+                console.log(public_id);
+
+                return deleteFromCloudinary(public_id);
+            });
+            await Promise.all(deletePromises);
+        }
+        if (bookData.images && bookData.images.length > 0) {
+            if (Array.isArray(bookData.images)) {
+                imageUrls.push(...bookData.images)
+            }
+            else {
+                imageUrls.push(bookData.images);
+            }
+        }
+        if (files && files.length != 0) {
+            const uploadPromises = files.map(file => uploadFromBuffer(file));
+            imageUrls.push(...(await Promise.all(uploadPromises)));
+        }
+
+        var dataToSave = {
+            ...bookData,
+            images: imageUrls.length == 0 ? undefined : imageUrls
+        }
+
+        if (bookData.categoryIds) {
+            const categories = await Promise.all(
+                bookData.categoryIds.map(id => this.categoryRepository.findById(id))
+            );
+            dataToSave.categories = categories;
+        }
+
+        if (bookData.authorIds) {
+            const authors = await Promise.all(
+                bookData.authorIds.map(id => this.authorRepository.findById(id))
+            );
+            dataToSave.authors = authors;
+        }
+        if (bookData.publisherId) {
+            var publisher = await this.publisherRepository.findById(bookData.publisherId);
+            dataToSave.publisher = publisher;
+        }
+
+        var newBook = await this.bookRepository.create(dataToSave);
+        return newBook;
+    }
+
+
+    async findAll({ page = 1, limit = 10, name = "" }) {
+        const books = await this.bookRepository.findAll({ page: page, limit: limit, name: name });
         return books;
     }
 
     async delete(id) {
+        const loanDetails = await this.loanDetailRepository.findByBookCopyIdOrBookId({ bookId: id })
+        console.log("loandetail", loanDetails)
+        if (loanDetails.length != 0) {
+            throw new ApiError(400, "Book is used, cannot delete")
+        }
         const book = await this.bookRepository.delete(id);
         await this.bookCopyService.deleteMany(book._id);
         const deletePromises = book.images.map(element => {
@@ -90,7 +137,6 @@ class BookService {
             return deleteFromCloudinary(public_id);
         });
         await Promise.all(deletePromises);
-        console.log("xoa", id);
         return book
     }
