+++ app/services/bookCopy.service.js
@@ -1,28 +1,28 @@
 import BookCopyRepository from "../repositories/bookCopy.repository.js";
+import LoanDetailRepository from "../repositories/loanDetail.repository.js";
 import MongoDB from "../utils/mongodb.util.js";
 
 class BookCopyService {
     constructor() {
         this.bookCopyRepository = new BookCopyRepository(MongoDB.client);
+        this.loanDetailRepository = new LoanDetailRepository(MongoDB.client);
     }
 
     async create(bookCopyRequest) {
-        if (bookCopyRequest.id && (bookCopyRequest.bookId == "" || bookCopyRequest.quantity == "" || bookCopyRequest.status == "")) {
-            throw new ApiError(400, "Params not valid")
-        }
         var bookCopies = []
-        if (!bookCopyRequest.id) {
-            var bookCopyCount = await this.bookCopyRepository.countByBookId(bookCopyRequest.bookId);
-            for (let i = 0; i < bookCopyRequest.quantity; i++) {
-                var barCode = bookCopyRequest.bookId + "-" + (bookCopyCount + 1);
-                bookCopyRequest.barCode = barCode;
-                bookCopyCount++;
-                bookCopies.push(await this.bookCopyRepository.create(bookCopyRequest));
-            }
+        var bookCopyCount = await this.bookCopyRepository.countByBookId(bookCopyRequest.bookId);
+        for (let i = 0; i < bookCopyRequest.quantity; i++) {
+            var barCode = bookCopyRequest.bookId + "-" + (bookCopyCount + 1);
+            bookCopyRequest.barCode = barCode;
+            bookCopyCount++;
+            bookCopies.push(await this.bookCopyRepository.create(bookCopyRequest));
         }
-        else
-            bookCopies.push(await this.bookCopyRepository.create(bookCopyRequest))
         return bookCopies;
+    }
+
+    async update(bookCopyRequest) {
+        const bookCopy = await this.bookCopyRepository.create(bookCopyRequest)
+        return bookCopy;
     }
 
     async findAll({ bookId, page = 1, limit = 10, status }) {
@@ -31,6 +31,10 @@ class BookCopyService {
     }
 
     async delete(id) {
+        const loanDetails = await this.loanDetailRepository.findByBookCopyIdOrBookId({ bookCopyId: id })
+        if (loanDetails.length != 0) {
+            throw new ApiError(400, "Book copy is used, cannot delete")
+        }
         const book = await this.bookCopyRepository.delete(id);
         return book;
     }
